# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Otis Financial Life Map API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List all loans
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanMutationResult"
...
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get a loan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanMutationResult"
        "404":
          description: Not found
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /loans/{id}/amortization:
    get:
      operationId: getLoanAmortization
      tags: [loans]
      summary: Get the full amortization schedule for a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Amortization schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanAmortization"
        "404":
          description: Not found

  # ── Forecasted Transactions ──────────────────────────────────────────────
...
  /retirement/settings:
    get:
      operationId: getRetirementSettings
      tags: [retirement]
      summary: Get retirement assumptions
      responses:
        "200":
          description: Retirement settings (defaults if never saved)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetirementSettings"
    post:
      operationId: saveRetirementSettings
      tags: [retirement]
      summary: Upsert retirement assumptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetirementSettingsInput"
      responses:
        "200":
          description: Saved retirement settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetirementSettings"
...
components:
  schemas:
...
    Loan:
      type: object
      required: [id, loanName, lenderName, loanType, originalAmount, currentBalance, interestRate, monthlyPayment, startDate, termMonths, nextPaymentDate, createdAt, updatedAt]
      properties:
        id:
          type: integer
        loanName:
          type: string
        lenderName:
          type: string
        loanType:
          type: string
        originalAmount:
          type: number
        currentBalance:
          type: number
        interestRate:
          type: number
        monthlyPayment:
          type: number
        startDate:
          type: string
        termMonths:
          type: integer
        nextPaymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BillSync:
      type: object
      required: [matched, billName]
      properties:
        matched:
          type: boolean
        billName:
          type: string
    LoanMutationResult:
      allOf:
        - $ref: "#/components/schemas/Loan"
        - type: object
          properties:
            billSync:
              $ref: "#/components/schemas/BillSync"
    LoanInput:
      type: object
      required: [loanName, lenderName, loanType, originalAmount, currentBalance, interestRate, monthlyPayment, startDate, termMonths, nextPaymentDate]
      properties:
        loanName:
          type: string
        lenderName:
          type: string
        loanType:
          type: string
        originalAmount:
          type: number
        currentBalance:
          type: number
        interestRate:
          type: number
        monthlyPayment:
          type: number
        startDate:
          type: string
        termMonths:
          type: integer
        nextPaymentDate:
          type: string
        notes:
          type: ["string", "null"]
    LoanUpdate:
      type: object
      properties:
        loanName:
          type: string
        lenderName:
          type: string
        loanType:
          type: string
        originalAmount:
          type: number
        currentBalance:
          type: number
        interestRate:
          type: number
        monthlyPayment:
          type: number
        startDate:
          type: string
        termMonths:
          type: integer
        nextPaymentDate:
          type: string
        notes:
          type: ["string", "null"]
...
    LoanAmortizationEntry:
      type: object
      required: [paymentNumber, paymentDate, paymentAmount, principal, interest, remainingBalance]
      properties:
        paymentNumber:
          type: integer
        paymentDate:
          type: string
        paymentAmount:
          type: number
        principal:
          type: number
        interest:
          type: number
        remainingBalance:
          type: number
    LoanAmortization:
      type: object
      required: [loanId, totalInterest, totalPaid, payoffDate, numberOfPayments, schedule]
      properties:
        loanId:
          type: integer
        totalInterest:
          type: number
        totalPaid:
          type: number
        payoffDate:
          type: ["string", "null"]
        numberOfPayments:
          type: integer
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/LoanAmortizationEntry"

    # ── Forecasted Transactions ───────────────────────────────────────────────
...
    RetirementSettings:
      type: object
      required: [currentAge, retirementAge, retirementGoal, expectedReturnRate, inflationRate, monthlySpendingGoal, socialSecurityMonthly, retirementDurationYears]
      properties:
        currentAge:
          type: ["integer", "null"]
        retirementAge:
          type: integer
        retirementGoal:
          type: ["number", "null"]
        expectedReturnRate:
          type: number
        inflationRate:
          type: number
        monthlySpendingGoal:
          type: ["number", "null"]
        socialSecurityMonthly:
          type: ["number", "null"]
        retirementDurationYears:
          type: integer
    RetirementSettingsInput:
      type: object
      required: [currentAge, retirementAge, retirementGoal, expectedReturnRate, inflationRate, monthlySpendingGoal, retirementDurationYears]
      properties:
        currentAge:
          type: integer
          minimum: 1
          maximum: 100
        retirementAge:
          type: integer
          minimum: 1
          maximum: 100
        retirementGoal:
          type: number
          minimum: 0
        expectedReturnRate:
          type: number
          minimum: 0
          maximum: 100
        inflationRate:
          type: number
          minimum: 0
          maximum: 100
        monthlySpendingGoal:
          type: number
          minimum: 0
        socialSecurityMonthly:
          type: ["number", "null"]
          minimum: 0
        retirementDurationYears:
          type: integer
          minimum: 1
          maximum: 60
...