# lib/api-zod/src/generated/types/loanMutationResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Otis Financial Life Map API
 * OpenAPI spec version: 0.1.0
 */
import type { BillSync } from './billSync';
import type { Loan } from './loan';

export type LoanMutationResult = Loan & {
  billSync?: BillSync;
};